Migrated link inventory base class
